add command

+++ src/main/java/com/software/bank/service/ICreditLogic.java
@@ -4,9 +4,9 @@ import com.software.bank.service.exception.ServiceException;
 
 public interface ICreditLogic {
 	
-	void addCredit(String contractNumber, long summa_credit, int term, double rate) throws ServiceException;
+	void addCredit(String contractNumber, double summa_credit, int term, double rate) throws ServiceException;
 	
-	void addPayment(String contractNumber, long summa_debet) throws ServiceException;
+	void addPayment(String contractNumber, double summa_debet) throws ServiceException;
 	
 	long getMinimalPayment(String contractNumber);
